Fix Signs bug (Untested)
This bug should be fixed, but it's untested. Please test with the latest builds, and let me know if this fixes it or not. Thanks.
FOr those that are wondering what the bug is, it's basically where when you have a sign, you can join the queue with any signs, even if they aren't a 1vs1 sign.